+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v4.0.0
+## Temporarily remove async upload with tw-upload
+Until a robust solution is found to accept the correct url.
+We temporarily stop async upload, because it fails and stops a user from submitting the form.
+
 # v3.13.0
 ## Feature enhancements for camera upload component
 Added camera overlay